roles

+++ app/Http/Controllers/ApiUserController.php
@@ -7,44 +7,54 @@ use App\Http\Requests;
 use Restaurant\User;
 //use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ApiUserController extends Controller
 {
 
     public function index()
-    { //return User::all();
-        $request = request();
+    {
+        if(Auth::user()->can('read-users')) {
+            $request = request();
 
-        // handle sort option
-        if (request()->has('sort')) {
-            list($sortCol, $sortDir) = explode('|', request()->sort);
-            $query = User::orderBy($sortCol, $sortDir);
-        } else {
-            $query = User::orderBy('id', 'asc');
+            // handle sort option
+            if (request()->has('sort')) {
+                list($sortCol, $sortDir) = explode('|', request()->sort);
+                $query = User::orderBy($sortCol, $sortDir);
+            } else {
+                $query = User::orderBy('id', 'asc');
+            }
+
+            if ($request->exists('filter')) {
+                $query->where(function($q) use($request) {
+                    $value = "%{$request->filter}%";
+                    $q->where('name', 'like', $value)
+                        ->orWhere('nickname', 'like', $value)
+                        ->orWhere('email', 'like', $value);
+                });
+            }
+
+            $perPage = request()->has('per_page') ? (int) request()->per_page : null;
+
+            // The headers 'Access-Control-Allow-Origin' and 'Access-Control-Allow-Methods'
+            // are to allow you to call this from any domain (see CORS for more info).
+            // This is for local testing only. You should not do this in production server,
+            // unless you know what it means.
+            return response()->json(
+                $query->paginate($perPage)
+            )
+            ->header('Access-Control-Allow-Origin', '*')
+            ->header('Access-Control-Allow-Methods', 'GET');
         }
-
-        if ($request->exists('filter')) {
-            $query->where(function($q) use($request) {
-                $value = "%{$request->filter}%";
-                $q->where('name', 'like', $value)
-                    ->orWhere('nickname', 'like', $value)
-                    ->orWhere('email', 'like', $value);
-            });
-        }
-
-        $perPage = request()->has('per_page') ? (int) request()->per_page : null;
-
-        // The headers 'Access-Control-Allow-Origin' and 'Access-Control-Allow-Methods'
-        // are to allow you to call this from any domain (see CORS for more info).
-        // This is for local testing only. You should not do this in production server,
-        // unless you know what it means.
-        return response()->json(
-            $query->paginate($perPage)
-        )
-        ->header('Access-Control-Allow-Origin', '*')
-        ->header('Access-Control-Allow-Methods', 'GET');
+        return redirect('/logout');
     }
-
+    public function create() {
+        if(Auth::user()->can('create-users')) {
+            return Role::orderBy('display_name', 'asc')->lists('display_name', 'id');
+            //return view('auth::user.create', compact('roles'));
+        }
+        return redirect('/logout');
+    }
     public function store(Request $request)
     {
         
@@ -67,9 +77,30 @@ class ApiUserController extends Controller
         return redirect()->route('users.index')
                         ->with('success','User created successfully');
                         */
-        return User::create($request->all());
+        if(Auth::user()->can('create-users')) {
+            return User::create($request->all());
+        }
+        return redirect('/logout');
+    }
+    /**
+     * muestra datos del recurso y combos asociados para editar
+     */
+    public function edit($id) {
+        if(Auth::user()->can('update-users')) {
+            $user = User::findOrFail($id);
+            $roles_user = User::find($id)->roles()->lists('role_id')->toArray();
+            $roles = Role::orderBy('display_name', 'asc')->lists('display_name', 'id');
+            return view('auth::user.edit', compact('user', 'roles', 'roles_user'));
+        }
+        return redirect('/logout');
     }
 
+    /*public function show() {
+        return view('auth::user.form_change_password');
+    }*/
+    /**
+     * solo muestra datos del recurso
+     */
     public function show($id)
     {
         /*
@@ -78,7 +109,10 @@ class ApiUserController extends Controller
         $userRole = $user->roles->lists('id','id')->toArray();
 
         return view('users.edit',compact('user','roles','userRole'));*/
-        return User::findOrFail($id);
+        if(Auth::user()->can('read-users')) {
+            return User::findOrFail($id);
+        }
+        return redirect('/logout');
     }
 
     public function update(Request $request, $id)
@@ -110,13 +144,19 @@ class ApiUserController extends Controller
         return redirect()->route('users.index')
                         ->with('success','User updated successfully');
                         */
-        User::findOrFail($id)->update($request->all());
-        return response()->json($request->all()); //response()->json()
-        
+        if(Auth::user()->can('update-users')) {
+
+            User::findOrFail($id)->update($request->all());
+            return response()->json($request->all()); //response()->json()
+        }
+        return redirect('/logout');
     }
 
     public function destroy($id)
     {
-        return User::destroy($id);
+        if($this->user->can('delete-users')) {
+            return User::destroy($id);
+        }
+        return redirect('/logout');
     }
 }
